% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_img.R
\name{chat_img}
\alias{chat_img}
\title{Function to upload image file}
\usage{
chat_img(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID"),
  file_path,
  message = "Image file uploaded."
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to post the text and R code}

\item{file_path}{file path of the image file you want to upload}

\item{file_message}{the body of the message you post with the file}
}
\description{
This function is used to upload image file to the specified roomid.
}
\examples{
chat_img(file_message = "Hello Chatwork image.",
         file_path = "path_to_image_file")
}
