% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_contacts.R
\name{chatr_contacts}
\alias{chatr_contacts}
\title{Function to get my user contacts}
\usage{
chatr_contacts(api_token = Sys.getenv("CHATWORK_API_TOKEN"))
}
\arguments{
\item{api_token}{your full ChatWork API token}
}
\description{
This function is used to access the list of users you are in contact with.
}
\examples{
chatr_contacts()
}
