% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_me.R
\name{chatr_me}
\alias{chatr_me}
\title{Function to get my account information}
\usage{
chatr_me(api_token = Sys.getenv("CHATWORK_API_TOKEN"))
}
\arguments{
\item{api_token}{your full ChatWork API token}
}
\description{
This function is used to send a text or an R code to the specified roomid.
}
\examples{
chatr_me()
}
