% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_my_status.R
\name{chatr_my_status}
\alias{chatr_my_status}
\title{Function to get my status}
\usage{
chatr_my_status(api_token = Sys.getenv("CHATWORK_API_TOKEN"))
}
\arguments{
\item{api_token}{your full ChatWork API token}
}
\description{
This function is used to get my unread count, unread TO count, and unfinished task count
}
\examples{
chatr_my_status()
}
