% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_delete.R
\name{chatr_room_delete}
\alias{chatr_room_delete}
\title{Function to leave or delete chat room}
\usage{
chatr_room_delete(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID"),
  action_type = c("leave", "delete"),
  confirm = FALSE
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to update}

\item{action_type}{leave or delete}

\item{confirm}{check if you really want to leave or delete. default is FALSE.}
}
\description{
This function is used to leave from chat room or delete chat room.
}
\examples{
chatr_room_delete(action_type = "leave", confirm = TRUE)
}
