% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_delete_message.R
\name{chatr_room_delete_message}
\alias{chatr_room_delete_message}
\title{Function to delete specified chat room message}
\usage{
chatr_room_delete_message(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID"),
  message_id = NULL,
  confirm = FALSE
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to get}

\item{message_id}{which message to delete}

\item{confirm}{check if you really want to delete}
}
\description{
This function is used to delete specified chat room message
}
\examples{
chatr_room_delete_message(message_id = "111111111111111")
}
