% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_get.R
\name{chatr_room_get}
\alias{chatr_room_get}
\title{Function to get specified chat room informations}
\usage{
chatr_room_get(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID")
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to get}
}
\description{
This function is used to get specified chat room informations.
}
\examples{
chatr_room_get()
}
