% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_get_file.R
\name{chatr_room_get_file}
\alias{chatr_room_get_file}
\title{Function to get specified chat room specified file}
\usage{
chatr_room_get_file(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID"),
  file_id = NULL,
  create_download_url = 0
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to get}

\item{file_id}{which file_id to get}

\item{create_download_url}{whether to generate a URL to download(0 is No, 1 is Yes).}
}
\description{
This function is used to get specified chat room specified file.
}
\examples{
chatr_room_get_file(file_id = "111111111", create_download_url = 1)
}
