% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_get_members.R
\name{chatr_room_get_members}
\alias{chatr_room_get_members}
\title{Function to get specified chat room members}
\usage{
chatr_room_get_members(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID")
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to get chat room informations.}
}
\description{
This function is used to get specified chat room members.
}
\examples{
chatr_room_get_members()
}
