% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_get_message.R
\name{chatr_room_get_message}
\alias{chatr_room_get_message}
\title{Function to get specified chat room message}
\usage{
chatr_room_get_message(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID"),
  message_id = NULL
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to get}

\item{message_id}{which message to get}
}
\description{
This function is used to get specified chat room message
}
\examples{
chatr_room_get_message(message_id = "111111111111111")
}
