% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_get_messages.R
\name{chatr_room_get_messages}
\alias{chatr_room_get_messages}
\title{Function to get specified chat room messages}
\usage{
chatr_room_get_messages(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID"),
  force = 0
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to get chat room informations.}

\item{force}{default is 0. If 1 is specified, the latest 100 items will be acquired regardless of whether they have not been acquired.}
}
\description{
This function is used to get specified chat room messages
}
\examples{
chatr_room_get_messages(force = 0)
}
