% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_get_task.R
\name{chatr_room_get_task}
\alias{chatr_room_get_task}
\title{Function to get specified chat room specified task}
\usage{
chatr_room_get_task(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID"),
  task_id = NULL
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to get}

\item{task_id}{which task to get}
}
\description{
This function is used to get specified chat room specified task
}
\examples{
chatr_room_get_task(task_id = "11111")
}
