% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_get_tasks.R
\name{chatr_room_get_tasks}
\alias{chatr_room_get_tasks}
\title{Function to get specified chat room tasks}
\usage{
chatr_room_get_tasks(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID"),
  account_id = NULL,
  assigned_by_account_id = NULL,
  status = c("open", "done")
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to get}

\item{account_id}{account id of the person in charge of the task}

\item{assigned_by_account_id}{account id of the task requester}

\item{status}{default is "open". If "done" is specified, Completed tasks will be acquired.}
}
\description{
This function is used to get specified chat room tasks
}
\examples{
chatr_room_get_tasks(account_id = "11111", assigned_by_account_id = "11111", status = "open")
}
