% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_post.R
\name{chatr_room_post}
\alias{chatr_room_post}
\title{Function to create chat room}
\usage{
chatr_room_post(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  description = "This room description is created by chatr through API.",
  icon_preset = "meeting",
  link = 0,
  link_need_acceptance = 0,
  members_admin_ids = NULL,
  members_member_ids = NULL,
  members_readonly_ids = NULL,
  name = "This room name is created by chatr through API."
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{description}{group chat overview text. default is "This room description is created by chatr through API.".}

\item{icon_preset}{group chat icon type. default is "meeting". choose from
{group, check, document, meeting, event, project, business, study, security, star, idea, heart, magcup, beer, music, sports, travel}.}

\item{link}{whether to create an invitation link. default is 0.}

\item{link_need_acceptance}{whether to participate by administrator approval. default is 0.}

\item{members_admin_ids}{character vactor of account IDs of users who want to have administrator privileges among the members who participate in the created chat.
at least one person must be specified.
you can only specify the account IDs of contacted users or users in your organization.}

\item{members_member_ids}{character vector of account IDs of the users who want to have member privileges among the members who participate in the created chat.
you can only specify the account IDs of contacted users or users in your organization.}

\item{members_readonly_ids}{character vaector of account IDs of the members who participate in the created chat and want to have read-only permission.
you can only specify the account IDs of contacted users or users in your organization.}

\item{name}{chat name of the group chat you want to create. default is "This room name is created by chatr through API.".}
}
\description{
This function is used to create chat room.
}
\examples{
chatr_room_post(
    description = "For Project X you can use it.",
    icon_preset = "meeting",
    link = 0,
    link_need_acceptance = 0,
    members_admin_ids = c("11111111","22222222"),
    members_member_ids = c("33333333"),
    members_readonly_ids = NULL,
    name = "Project X"
    )
}
