% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_post_tasks.R
\name{chatr_room_post_tasks}
\alias{chatr_room_post_tasks}
\title{Function to create specified chat room tasks}
\usage{
chatr_room_post_tasks(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID"),
  body = NULL,
  limit = NULL,
  limit_type = c("none", "date", "time"),
  to_ids = NULL
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to get}

\item{body}{task message}

\item{limit}{task deadline specified in datetime. format is `yyyy-mm-dd hh:mm:ss`.}

\item{limit_type}{task deadline type(none, date, time)}

\item{to_ids}{account ids of the person in charge}
}
\description{
This function is used to create specified chat room tasks
}
\examples{
chatr_room_post_tasks(body = "sample task",
                      limit = "2020-12-15 00:00:00", # unix time is 16079583600
                      limit_type = "time",
                      to_ids = c("1111111","222222"))
}
