% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_put.R
\name{chatr_room_put}
\alias{chatr_room_put}
\title{Function to create chat room}
\usage{
chatr_room_put(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID"),
  description = NULL,
  icon_preset = NULL,
  name = NULL
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to update}

\item{description}{group chat overview text you want to update.}

\item{icon_preset}{group chat icon type you want to update. choose from
{group, check, document, meeting, event, project, business, study, security, star, idea, heart, magcup, beer, music, sports, travel}.}

\item{name}{chat name of the group chat you want to update}
}
\description{
This function is used to create chat room.
}
\examples{
chatr_room_put(
    description = "description you want to update",
    name = "name you want to update"
    )
}
