% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_put_members.R
\name{chatr_room_put_members}
\alias{chatr_room_put_members}
\title{Function to update specified chat room authority}
\usage{
chatr_room_put_members(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID"),
  members_admin_ids = NULL,
  members_member_ids = NULL,
  members_readonly_ids = NULL
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to update}

\item{members_admin_ids}{character vactor of account IDs of users who want to have administrator privileges among the members who participate in the created chat.
At least one id must be specified. You can only specify the account IDs of contacted users or users in your organization.}

\item{members_member_ids}{character vactor of account IDs of users who want to have member privileges among the members who participate in the created chat.
You can only specify the account IDs of contacted users or users in your organization.}

\item{members_readonly_ids}{character vactor of account IDs of users who want to have readonly privileges among the members who participate in the created chat.
You can only specify the account IDs of contacted users or users in your organization.}
}
\description{
This function is used to update specified chat room authority
}
\examples{
chatr_room_put_members(
    members_admin_ids = c("11111111","22222222"),
    members_member_ids = c("33333333"),
    members_readonly_ids = NULL,
    )
}
