% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_put_message.R
\name{chatr_room_put_message}
\alias{chatr_room_put_message}
\title{Function to update specified chat room message}
\usage{
chatr_room_put_message(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID"),
  message_id = NULL,
  body = NULL
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to get}

\item{message_id}{which message to update}

\item{body}{text you want to update}
}
\description{
This function is used to update specified chat room message
}
\examples{
chatr_room_put_message(message_id = "111111111111111", body = "This is an updated text")
}
