% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_room_put_task.R
\name{chatr_room_put_task}
\alias{chatr_room_put_task}
\title{Function to update specified chat room task status}
\usage{
chatr_room_put_task(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  roomid = Sys.getenv("CHATWORK_ROOMID"),
  task_id = NULL,
  body = c("open", "done")
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{roomid}{which room to get}

\item{task_id}{which task to update}

\item{body}{which status you change(open or done)}
}
\description{
This function is used to update specified chat room task status
}
\examples{
chatr_room_put_task(task_id = "11111111", body = "done")
}
