% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{chatR_server}
\alias{chatR_server}
\title{The Shiny server function for the custom chatbot.}
\usage{
chatR_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{session}{Shiny session object.}

\item{outout}{Shiny output object.}
}
\description{
The Shiny server function for the custom chatbot.
}
