% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_tasks.R
\name{chatr_tasks}
\alias{chatr_tasks}
\title{Function to get my tasks}
\usage{
chatr_tasks(
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  assigned_by_account_id = NULL,
  status = "open"
)
}
\arguments{
\item{api_token}{your full ChatWork API token}

\item{assigned_by_account_id}{person id who assigned the task}

\item{status}{select open task or closed task}
}
\description{
This function is used to get my task list.
}
\examples{
chatr_tasks()
}
