% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{chatR_ui}
\alias{chatR_ui}
\title{The Shiny UI for the custom chatbot.}
\usage{
chatR_ui()
}
\value{
a \code{\link[shiny:navbarPage]{shiny::navbarPage()}} object.
}
\description{
The Shiny UI for the custom chatbot.
}
