% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conv_datetime2unixtime}
\alias{conv_datetime2unixtime}
\title{Unix time converter to datetime}
\usage{
conv_datetime2unixtime(datetime)
}
\arguments{
\item{datetime}{datetime}
}
\description{
This function is used to convert from JST datetime to Unix time
}
\examples{
conv_datetime2unixtime("2020-12-12 00:00:00")
}
