% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conv_datetime_to_unixtime}
\alias{conv_datetime_to_unixtime}
\title{Unix time converter to datetime}
\usage{
conv_datetime_to_unixtime(datetime)
}
\arguments{
\item{datetime}{datetime}
}
\description{
This function is used to convert from JST datetime to Unix time
}
\examples{
datetime <- "2020-12-12 15:23:17"
conv_datetime_to_unixtime(datetime)
}
