% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_chatroom_membership}
\alias{create_chatroom_membership}
\title{Create a chatroom membership}
\usage{
create_chatroom_membership(bearer_token, user_id, chatroom_id, instruction_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{user_id}{A numeric. The unique ID for the user.}

\item{chatroom_id}{A numeric. The unique ID for the chatroom.}

\item{instruction_id}{A numeric. The unique ID for the instruction.}
}
\value{
Returns ...
}
\description{
Create a chatroom membership
}
\examples{

\dontrun{
#create experiment, instruction, user, and chatroom first
create_chatroom_membership(bearer_token = "<your-token-here>",
                           user_id = 1,
                           chatroom_id = 1,
                           instruction_id = 1)
}

}
