% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_user}
\alias{create_user}
\title{Create a user}
\usage{
create_user(bearer_token, username)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{username}{A string. Unique identifier for the user. Only visible to researcher.}
}
\value{
Returns ...
}
\description{
Create a user
}
\examples{

\dontrun{
create_user(bearer_token = "<your-token-here>",
            username = "My First User")
}

}
