% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_chatroom}
\alias{delete_chatroom}
\title{Delete a chatroom}
\usage{
delete_chatroom(bearer_token, chatroom_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{chatroom_id}{A numeric. The unique ID for the chatroom.}
}
\value{
Returns ...
}
\description{
Delete a chatroom
}
\examples{

\dontrun{
delete_chatroom(bearer_token = "<your-token-here>")
}

}
