% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_chatroom_membership}
\alias{delete_chatroom_membership}
\title{Delete a chatroom membership}
\usage{
delete_chatroom_membership(bearer_token, chatroom_membership_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{chatroom_membership_id}{A numeric. The unique ID for the chatroom membership.}
}
\value{
Returns ...
}
\description{
Delete a chatroom membership
}
\examples{

\dontrun{
delete_chatroom_membership(bearer_token = "<your-token-here>")
}

}
