% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_experiment}
\alias{delete_experiment}
\title{Delete an experiment}
\usage{
delete_experiment(bearer_token, experiment_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{experiment_id}{A numeric. The unique ID for the experiment.}
}
\value{
Returns ...
}
\description{
Delete an experiment
}
\examples{

\dontrun{
delete_experiment(bearer_token = "<your-token-here>")
}

}
