% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_message}
\alias{delete_message}
\title{Delete a message}
\usage{
delete_message(bearer_token, message_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{message_id}{A numeric. The unique ID for the message.}
}
\value{
Returns ...
}
\description{
Delete a message
}
\examples{

\dontrun{
delete_message(bearer_token = "<your-token-here>")
}

}
