% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_user}
\alias{delete_user}
\title{Delete a user}
\usage{
delete_user(bearer_token, user_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{user_id}{A numeric. The unique ID for the user.}
}
\value{
Returns ...
}
\description{
Delete a user
}
\examples{

\dontrun{
delete_user(bearer_token = "<your-token-here>")
}

}
