% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_datetime_format}
\alias{is_datetime_format}
\title{Check task deadline datetime format}
\usage{
is_datetime_format(datetime)
}
\arguments{
\item{datetime}{datetime}
}
\description{
This function is used to check datetime format
}
\examples{
is_datetime_format("2020-12-12 00:00:00")
}
