% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_deadline}
\alias{is_deadline}
\title{Check task deadline datetime}
\usage{
is_deadline(datetime)
}
\arguments{
\item{datetime}{datetime}
}
\description{
This function is used to check task deadline date
}
\examples{
is_deadline(datetime)
}
