% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_file_limit}
\alias{is_file_limit}
\title{Chack file size}
\usage{
is_file_limit(file_path)
}
\arguments{
\item{file_path}{file path of the image or video file you want to upload}
}
\description{
This function is used to chack file size.
}
\examples{
is_file_limit(file_path = ~/path_to_file)
}
