% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_room_icons}
\alias{is_room_icons}
\title{Check room icon}
\usage{
is_room_icons(icon_preset)
}
\arguments{
\item{icon_preset}{icon you want to update}
}
\description{
This function is used to check room icon.
}
\examples{
is_room_icons(icon_preset)
}
