% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_chatroom_memberships}
\alias{list_chatroom_memberships}
\title{List all chatroom memberships}
\usage{
list_chatroom_memberships(bearer_token)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}
}
\value{
Returns ...
}
\description{
List all chatroom memberships
}
\examples{

\dontrun{
list_chatroom_memberships(bearer_token = "<your-token-here>")
}

}
