% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_chatrooms}
\alias{list_chatrooms}
\title{List all chatrooms}
\usage{
list_chatrooms(bearer_token)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
List all chatrooms
}
\examples{

\dontrun{
list_chatrooms(bearer_token = "<your-token-here>")
}

}
