% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_experiments}
\alias{list_experiments}
\title{List all experiments}
\usage{
list_experiments(bearer_token)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}
}
\value{
Returns ...
}
\description{
List all experiments
}
\examples{

\dontrun{
list_experiments(bearer_token = "<your-token-here>")
}

}
