% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_users}
\alias{list_users}
\title{List all users}
\usage{
list_users()
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
List all users
}
\examples{

\dontrun{
list_users()
}

}
