% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Climate models upload}
\usage{
load_data(
  path.to.rcps,
  country,
  variable,
  xlim = NULL,
  ylim = NULL,
  years = 2010:2099,
  path.to.obs = NULL,
  n.cores,
  buffer = 0
)
}
\arguments{
\item{path.to.rcps}{Absolute path to the directory containing the RCPs/SSPs folders and historical simulations. For example,
home/user/data/. data would contain subfolders with the climate models. Historical simulations have to be contained in a folder called historical}

\item{country}{A character string, in english, indicating the country of interest. To select a bounding box,
set country to NULL and define arguments xlim and ylim}

\item{variable}{A character string indicating the variable to be loaded}

\item{xlim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box selected.}

\item{ylim}{Same as xlim, but for the selection of the latitudinal range.}

\item{years}{Numerical range, years to select. Default (NULL)}

\item{path.to.obs}{Default to NULL, if not, indicate the absolute path to the directory containing a reanalysis dataset, for example W5E5 or ERA5.}

\item{n.cores}{Integer, number of cores to use in parallel processing}

\item{buffer}{Integer, default to zero. Buffer to add when selecting a country or a bounding box}
}
\value{
Tibble with list column
}
\description{
Automatically load climate models (netCDF/NcML) in a tidy format.
}
\examples{
fpath <- system.file("extdata/", package="chatR")
exmp <- load_data(country = "Moldova", variable="hurs", n.cores=6,
              path.to.rcps = fpath)
}
