% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recruit_ex}
\alias{recruit_ex}
\title{Simulated recruitment data}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{id}{unique identifer}
\item{pid}{partisanship indicator}
}
}
\usage{
recruit_ex
}
\description{
An example dataset containing a fake id variable and binary
partisanship indicator for 100 observations.
}
\keyword{datasets}
