% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_chatroom_membership}
\alias{retrieve_chatroom_membership}
\title{Retrieve a chatroom membership}
\usage{
retrieve_chatroom_membership(bearer_token, chatroom_membership_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{chatroom_membership_id}{A numeric. The unique ID for the chatroom membership.}
}
\value{
Returns ...
}
\description{
Retrieve a chatroom membership
}
\examples{

\dontrun{
retrieve_chatroom_membership(bearer_token = "<your-token-here>")
}

}
