% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_experiment}
\alias{retrieve_experiment}
\title{Retrieve an experiment}
\usage{
retrieve_experiment(bearer_token, experiment_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{experiment_id}{A numeric. The unique ID for the experiment.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Retrieve an experiment
}
\examples{

\dontrun{
retrieve_experiment(bearer_token = "<your-token-here>")
}

}
