% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_instruction}
\alias{retrieve_instruction}
\title{Retrieve an instruction}
\usage{
retrieve_instruction(instruction_id)
}
\arguments{
\item{instruction_id}{A numeric. The unique ID for the instruction.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Retrieve an instruction
}
\examples{

\dontrun{
retrieve_instruction(instruction_id = 1)
}

}
