% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_message}
\alias{retrieve_message}
\title{Retrieve a message}
\usage{
retrieve_message(message_id)
}
\arguments{
\item{message_id}{A numeric. The unique ID for the message.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Retrieve a message
}
\examples{

\dontrun{
retrieve_message(message_id = 1)
}

}
