% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_user}
\alias{retrieve_user}
\title{Retrieve a user}
\usage{
retrieve_user(bearer_token, user_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{user_id}{A numeric. The unique ID for the user.}
}
\value{
Returns ...
}
\description{
Retrieve a user
}
\examples{

\dontrun{
retrieve_user(bearer_token = "<your-token-here>")
}

}
