% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chatR.R
\name{run_chatR}
\alias{run_chatR}
\title{Run the chatR Shiny application}
\usage{
run_chatR(
  config_file,
  store_location = "",
  system_prompt = "",
  model = "llama3.1",
  base_url = Sys.getenv("OLLAMA_BASE_URL", "http://localhost:11434"),
  history_dir = "history",
  app_title = "chatR"
)
}
\arguments{
\item{config_file}{a configuratino file. If specified, all other parameters are ignored.}
}
\description{
Run the chatR Shiny application
}
