% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sign.prop}
\alias{sign.prop}
\title{Climate change signal agreement}
\usage{
\method{sign}{prop}(array3d)
}
\description{
function to find models members sign agreement. Input is a 3d array with first dimension being members
}
