% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{throw_error}
\alias{throw_error}
\title{display API request error}
\usage{
throw_error(response)
}
\description{
This function is used to catch and throw API request error
}
\examples{
throw_error(response)
}
