% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_chatroom}
\alias{update_chatroom}
\title{Update a chatroom}
\usage{
update_chatroom(
  bearer_token,
  chatroom_id,
  topic = NULL,
  min_duration = NULL,
  max_duration = NULL
)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{chatroom_id}{A numeric. The unique ID for the chatroom.}

\item{topic}{Optional. A string. Unique identifier for the chatroom. Only visible to researcher.}

\item{min_duration}{Optional. A numeric.  The minimum time in seconds users must spend in the chatroom before the "Done" button is active.}

\item{max_duration}{Optional. A numeric.  The maximum time in seconds users must spend in the chatroom before the "Done" button is active.}
}
\value{
Returns ...
}
\description{
Update a chatroom
}
\examples{

\dontrun{
update_chatroom(bearer_token = "<your-token-here>",
                chatroom_id = "1",
                max_duration = 900)
}

}
