% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_chatroom_membership}
\alias{update_chatroom_membership}
\title{Update a chatroom membership}
\usage{
update_chatroom_membership(
  bearer_token,
  chatroom_membership_id,
  user_id = NULL,
  chatroom_id = NULL,
  instruction_id = NULL
)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{chatroom_membership_id}{A numeric. The unique ID for the chatroom membership.}

\item{user_id}{Optional. A numeric. The unique ID for the user.}

\item{chatroom_id}{Optional. A numeric. The unique ID for the chatroom.}

\item{instruction_id}{Optional. A numeric. The unique ID for the instruction.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the updated content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Update a chatroom membership
}
\examples{

\dontrun{
update_chatroom_membership(bearer_token = "<your-token-here>",
                           chatroom_membership_id = 1,
                           instruction_id = 2) # Change instructions
}

}
