% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_experiment}
\alias{update_experiment}
\title{Update an experiment}
\usage{
update_experiment(
  bearer_token,
  experiment_id,
  name = NULL,
  post_survey_link = NULL,
  moderator_message = NULL,
  language = NULL
)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{experiment_id}{A numeric. The unique ID for the experiment.}

\item{name}{Optional. A string. The experiment's name.}

\item{post_survey_link}{Optional. A string. A link to direct users to after the chat}

\item{moderator_message}{Optional. A string. The first message in the chat labeled as the "moderator"}

\item{language}{Optional. A string. Language of the Chatter features. Defaults to "English" option for "Spanish".}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the updated content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Update an experiment
}
\examples{

\dontrun{
update_experiment(bearer_token = "<your-token-here>",
                  experiment_id = 1,
                  name = "New Name for Experiment")
}

}
