% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_instruction}
\alias{update_instruction}
\title{Update an instruction}
\usage{
update_instruction(
  bearer_token,
  instruction_id,
  experiment_id = NULL,
  name = NULL,
  text = NULL
)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{instruction_id}{A numeric. A numeric. The unique ID for the instruction.}

\item{experiment_id}{Optional. A numeric. The unique ID for the experiment.}

\item{name}{Optional. A string. The instruction's name. Visible only to the researcher}

\item{text}{Optional. A string.  The instruction's contents.  Visible to the user.  HTML compatible.}
}
\value{
Returns ...
}
\description{
Update an instruction
}
\examples{

\dontrun{
update_instruction(bearer_token = "<your-token-here>",
                   instruction_id = 1,
                   name = "New Name for Instruction")
}

}
