% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_message}
\alias{update_message}
\title{Update a message}
\usage{
update_message(message_id, user_id = NULL, chatroom_id = NULL, content = NULL)
}
\arguments{
\item{message_id}{A numeric. The unique ID for the message.}

\item{user_id}{Optional. A numeric. The unique ID for the user. Note that the user posting the message does not have to be a member of the chatroom.}

\item{chatroom_id}{Optional. A numeric. The unique ID for the chatroom.}

\item{content}{Optional. A string.  The message's content.  Visible to all members of the chatroom.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the updated content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Update a message
}
\examples{

\dontrun{
update_message(message_id = 1,
               content = "updated message content")
}

}
