% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr.R
\name{chatr}
\alias{chatr}
\title{Send txet and R Script to specified chat room.}
\usage{
chatr(
  ...,
  code = FALSE,
  api_token = Sys.getenv("CHATWORK_API_TOKEN"),
  room_id = Sys.getenv("CHATWORK_ROOMID"),
  self_unread = 0
)
}
\arguments{
\item{...}{expressions to be sent to chat room}

\item{code}{whether to enclose text and R Script in [code][/code] tags or not}

\item{api_token}{your full ChatWork API token}

\item{room_id}{which room to post the text and R code}

\item{self_unread}{If 1 is specified, the message added by you will be marked as unread by yourself. default is 0(mark as read).}
}
\description{
This function is used to send a text or an R Script to the specified chat room.
}
\examples{
chatr(code = TRUE,
set.seed(1989),
x <- rnorm(10),
y <- rnorm(10),
f <- lm(y ~ x),
summary(f))

chatr(code = FALSE, "Hello ChatWork!")
}
