% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatr_setup.R
\name{chatr_setup}
\alias{chatr_setup}
\title{Setup environment variables for ChatWork API connection}
\usage{
chatr_setup(config_file_path = NULL)
}
\arguments{
\item{config_file_path}{path to config.yml file}
}
\description{
This function sets api_token and roomid as environment variables.
}
\examples{
chatr_setup(config_file_path = "~/config.yml")
}
