% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_chatroom}
\alias{create_chatroom}
\title{Create a chatroom}
\usage{
create_chatroom(topic, min_duration = NULL, max_duration = NULL)
}
\arguments{
\item{topic}{A string. Unique identifier for the chatroom. Only visible to researcher.}

\item{min_duration}{A numeric.  The minimum time in seconds users must spend in the chatroom before the "Done" button is active.}

\item{max_duration}{A numeric.  The maximum time in seconds users must spend in the chatroom before the "Done" button is active.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the content created, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Create a chatroom
}
\examples{

\dontrun{
create_chatroom(topic = "My First Chatroom",
                min_duration = 30,
                max_duration = 600)
}

}
