% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_chatroom_membership}
\alias{create_chatroom_membership}
\title{Create a chatroom membership}
\usage{
create_chatroom_membership(
  user_id,
  chatroom_id,
  instruction_id,
  display_name = NULL
)
}
\arguments{
\item{user_id}{A numeric. The unique ID for the user.}

\item{chatroom_id}{A numeric. The unique ID for the chatroom.}

\item{instruction_id}{A numeric. The unique ID for the instruction.}

\item{display_name}{Optional. A string. The name other chatroom members see associated with the user's messages.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the content created, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Create a chatroom membership
}
\examples{

\dontrun{
#create experiment, instruction, user, and chatroom first
create_chatroom_membership(user_id = 1,
                           chatroom_id = 1,
                           instruction_id = 1)
}

}
