% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_experiment}
\alias{create_experiment}
\title{Create an experiment}
\usage{
create_experiment(
  bearer_token,
  name,
  post_survey_link = NULL,
  moderator_message = NULL,
  language = "English"
)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{name}{A string. The experiment's name.}

\item{post_survey_link}{Optional. A string. A link to direct users to after the chat}

\item{moderator_message}{Optional. A string. The first message in the chat labeled as the "moderator"}

\item{language}{Optional. A string. Language of the Chatter features. Defaults to "English" option for "Spanish".}
}
\value{
Returns ...
}
\description{
Create an experiment
}
\examples{

\dontrun{
create_experiment(bearer_token = "<your-token-here>",
                  name = "My First Experiment")
}

}
