% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_instruction}
\alias{create_instruction}
\title{Create an instruction}
\usage{
create_instruction(bearer_token, experiment_id, name, text)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{experiment_id}{A numeric. ...}

\item{name}{A string. The instruction's name, visible only to the researcher}

\item{text}{A string.  The instruction's contents visible to the user.  HTML compatible.}
}
\value{
Returns ...
}
\description{
Create an instruction
}
\examples{

\dontrun{
create_instruction(bearer_token = "<your-token-here>",
                    experiment_id = 1,
                    name = "Instructions for My First Experiment",
                    text = "Follow these instructions.")
}

}
