% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_instruction}
\alias{create_instruction}
\title{Create an instruction}
\usage{
create_instruction(bearer_token, experiment_id, name, text)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{experiment_id}{A numeric. The unique ID for the experiment.}

\item{name}{A string. The instruction's name. Visible only to the researcher}

\item{text}{A string.  The instruction's contents.  Visible to the user.  HTML compatible.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the content created, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Create an instruction
}
\examples{

\dontrun{
create_instruction(bearer_token = "<your-token-here>",
                   experiment_id = 1, # create experiment first
                   name = "Instructions for My First Experiment",
                   text = "Follow these instructions.")
}

}
