% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_user}
\alias{create_user}
\title{Create a user}
\usage{
create_user(bearer_token, username)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{username}{A string. Unique identifier for the user. Only visible to researcher.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the content created, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Create a user
}
\examples{

\dontrun{
create_user(bearer_token = "<your-token-here>",
            username = "My First User")
}

}
