% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_chatroom}
\alias{delete_chatroom}
\title{Delete a chatroom}
\usage{
delete_chatroom(bearer_token, chatroom_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{chatroom_id}{A numeric. The unique ID for the chatroom.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is \code{NA} as no information is created or requested in a DELETE call, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Delete a chatroom
}
\examples{

\dontrun{
delete_chatroom(bearer_token = "<your-token-here>")
}

}
