% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_chatroom_membership}
\alias{delete_chatroom_membership}
\title{Delete a chatroom membership}
\usage{
delete_chatroom_membership(chatroom_membership_id)
}
\arguments{
\item{chatroom_membership_id}{A numeric. The unique ID for the chatroom membership.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is \code{NA} as no information is created or requested in a DELETE call, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Delete a chatroom membership
}
\examples{

\dontrun{
delete_chatroom_membership(chatroom_membership_id = 1)
}

}
