% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_chatroom_membership}
\alias{delete_chatroom_membership}
\title{Delete a chatroom membership}
\usage{
delete_chatroom_membership(chatroom_membership_id)
}
\arguments{
\item{chatroom_membership_id}{A numeric. The unique ID for the chatroom membership or a vector of multiple IDs.}
}
\value{
Returns a list of two elements. \code{content} is \code{NA} as no information is created or requested in a DELETE call.
\code{meta_data} is a list of meta data associated with the API call including the following elements. \code{path} is a
string indicating the path for each deleted object, \code{status} is a string indicating whether the API call was successful
or not, and \code{error} is a string indicating any error message provided.
}
\description{
Delete a chatroom membership
}
\examples{

\dontrun{
delete_chatroom_membership(chatroom_membership_id = 1)
}

}
