% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_experiment}
\alias{delete_experiment}
\title{Delete an experiment}
\usage{
delete_experiment(experiment_id)
}
\arguments{
\item{experiment_id}{A numeric. The unique ID for the experiment.}
}
\value{
Returns a data.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is \code{NA} as no information is created or requested in a DELETE call, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Delete an experiment
}
\examples{

\dontrun{
delete_experiment(experiment_id = 1)
}

}
