% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_instruction}
\alias{delete_instruction}
\title{Delete an instruction}
\usage{
delete_instruction(instruction_id)
}
\arguments{
\item{instruction_id}{A numeric. The unique ID for the instruction.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is \code{NA} as no information is created or requested in a DELETE call, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Delete an instruction
}
\note{
Note you need to delete any chatroom memberships that contain the instruction
before you can successfully delete the instruction.
}
\examples{

\dontrun{
delete_instruction(instruction_id = 1)
}

}
