% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_instruction}
\alias{delete_instruction}
\title{Delete an instruction}
\usage{
delete_instruction(bearer_token, instruction_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{instruction_id}{A numeric. The unique ID for the instruction.}
}
\value{
Returns ...
}
\description{
Delete an instruction
}
\note{
Note you need to delete any chatroom memberships that contain the instruction
before you can successfully delete the instruction.
}
\examples{

\dontrun{
delete_instruction(bearer_token = "<your-token-here>")
}

}
