% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_message}
\alias{delete_message}
\title{Delete a message}
\usage{
delete_message(message_id)
}
\arguments{
\item{message_id}{A numeric. The unique ID for the message.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is \code{NA} as no information is created or requested in a DELETE call, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Delete a message
}
\examples{

\dontrun{
delete_message(message_id = 1)
}

}
