% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_user}
\alias{delete_user}
\title{Delete a user}
\usage{
delete_user(user_id)
}
\arguments{
\item{user_id}{A numeric. The unique ID for the user.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is \code{NA} as no information is created or requested in a DELETE call, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Delete a user
}
\examples{

\dontrun{
delete_user(user_id = 1)
}

}
