% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_chatrooms}
\alias{list_chatrooms}
\title{List all chatrooms}
\usage{
list_chatrooms(bearer_token)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}
}
\value{
Returns ...
}
\description{
List all chatrooms
}
\examples{

\dontrun{
list_chatrooms(bearer_token = "<your-token-here>")
}

}
