% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_experiments}
\alias{list_experiments}
\title{List all experiments}
\usage{
list_experiments()
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.

Returns a list of two elements. \code{content} is a dataframe of the requested content.
\code{meta_data} is a dataframe of meta data associated with the API call for each page of the \code{content} dataframe- \code{path} is a
string indicating the path for each page of data, \code{status} is a string indicating whether the API call was successful
or not, and \code{error} is a string indicating any error message provided.
}
\description{
List all experiments
}
\examples{

\dontrun{
list_experiments()
}

}
