% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_messages}
\alias{list_messages}
\title{List all messages}
\usage{
list_messages(bearer_token)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}
}
\value{
Returns ...
}
\description{
List all messages
}
\examples{

\dontrun{
list_messages(bearer_token = "<your-token-here>")
}

}
