% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_users}
\alias{list_users}
\title{List all users}
\usage{
list_users(bearer_token)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}
}
\value{
Returns ...
}
\description{
List all users
}
\examples{

\dontrun{
list_users(bearer_token = "<your-token-here>")
}

}
