% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_chatroom}
\alias{retrieve_chatroom}
\title{Retrieve a chatroom}
\usage{
retrieve_chatroom(chatroom_id)
}
\arguments{
\item{chatroom_id}{A numeric. The unique ID for the chatroom.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Retrieve a chatroom
}
\examples{

\dontrun{
retrieve_chatroom(chatroom_id = 1)
}

}
