% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_experiment}
\alias{retrieve_experiment}
\title{Create an experiment}
\usage{
retrieve_experiment(bearer_token, experiment_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{experiment_id}{asdf}
}
\value{
Returns ...
}
\description{
Create an experiment
}
\examples{

\dontrun{
create_experiment(bearer_token = "<your-token-here>",
                  name = "My First Experiment")
}

}
