% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_instruction}
\alias{retrieve_instruction}
\title{Retrieve an instruction}
\usage{
retrieve_instruction(bearer_token, instruction_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{instruction_id}{A numeric. The unique ID for the instruction.}
}
\value{
Returns ...
}
\description{
Retrieve an instruction
}
\examples{

\dontrun{
retrieve_instruction(bearer_token = "<your-token-here>")
}

}
