% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_message}
\alias{retrieve_message}
\title{Retrieve a message}
\usage{
retrieve_message(bearer_token, message_id)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{message_id}{A numeric. The unique ID for the message.}
}
\value{
Returns ...
}
\description{
Retrieve a message
}
\examples{

\dontrun{
retrieve_message(bearer_token = "<your-token-here>")
}

}
