% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_user}
\alias{retrieve_user}
\title{Retrieve a user}
\usage{
retrieve_user(user_id)
}
\arguments{
\item{user_id}{A numeric. The unique ID for the user.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the requested content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Retrieve a user
}
\examples{

\dontrun{
retrieve_user(user_id = 1)
}

}
