% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_instruction}
\alias{update_instruction}
\title{Update an instruction}
\usage{
update_instruction(
  instruction_id,
  experiment_id = NULL,
  name = NULL,
  text = NULL
)
}
\arguments{
\item{instruction_id}{A numeric. A numeric. The unique ID for the instruction.}

\item{experiment_id}{Optional. A numeric. The unique ID for the experiment.}

\item{name}{Optional. A string. The instruction's name. Visible only to the researcher}

\item{text}{Optional. A string.  The instruction's contents.  Visible to the user.  HTML compatible.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the updated content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Update an instruction
}
\examples{

\dontrun{
update_instruction(instruction_id = 1,
                   name = "New Name for Instruction")
}

}
