% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_message}
\alias{update_message}
\title{Update a message}
\usage{
update_message(
  bearer_token,
  message_id,
  user_id = NULL,
  chatroom_id = NULL,
  content = NULL
)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{message_id}{A numeric. The unique ID for the message.}

\item{user_id}{Optional. A numeric. The unique ID for the user. Note that the user posting the message does not have to be a member of the chatroom.}

\item{chatroom_id}{Optional. A numeric. The unique ID for the chatroom.}

\item{content}{Optional. A string.  The message's content.  Visible to all members of the chatroom.}
}
\value{
Returns ...
}
\description{
Update a message
}
\examples{

\dontrun{
update_message(bearer_token = "<your-token-here>",
               message_id = 1,
               content = "updated message content")
}

}
