% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_user}
\alias{update_user}
\title{Update a user}
\usage{
update_user(bearer_token, user_id, username = NULL)
}
\arguments{
\item{bearer_token}{A string. The researcher's Auth Token found in Chatter online interface > API Credentials.}

\item{user_id}{A numeric. The unique ID for the user.}

\item{username}{Optional. A string. Unique identifier for the user. Only visible to researcher.}
}
\value{
Returns a list of four elements.  \code{status} is a string indicating whether the API call was successful or not,
\code{error} is a string indicating any error message provided, \code{content} is a dataframe of the updated content, and
\code{meta_data} is a list of meta data associated with the API call.
}
\description{
Update a user
}
\examples{

\dontrun{
update_user(bearer_token = "<your-token-here>",
            user_id = 1,
            username = "New Username")
}

}
